requirements.txt
Appium-Python-Client
